Changing Read/WriteUint32/64At functions to native endianness
Also adding a Sync() method
The binary.LittleEndian approach was causing a panic for me
when writing to /dev/mem on a Xilinx Zynq system for AXI register access.
I'm guessing it's because the write was not a single operation.
Neither this approach nor the the former one completely satisfies all
possible endianness issues. But I'm a selfish developer with selfish
needs.